수정 후 코밋코밋 내 이름은 코밋모크코밋

+++ MySkin/src/main/java/kr/smhrd/domain/ProductsDTO.java
@@ -12,7 +12,7 @@ public class ProductsDTO {
 	private String prod_name;
 	private int prod_price;
 	private String prod_maker;
-	private int prod_rating;
+	private double prod_rating;
 	private String prod_ex_date;
 	
 	public String getFormat_price() {
